update to improved SmithWaterman interface
git-svn-id: file:///home/svn-repositories/biojava/biojava-live/trunk@4811 7c6358e6-4a41-0410-a743-a5b2a554c398

+++ src/org/biojava/bio/structure/io/SeqRes2AtomAligner.java
@@ -229,7 +229,7 @@ public class SeqRes2AtomAligner {
             Sequence bjseq2 = ProteinTools.createProteinSequence(seq2,"seq2");
 
             //System.out.println(bjseq1.getAlphabet());
-            NeedlemanWunsch aligner = new NeedlemanWunsch(-2, 5, 3, 3, 0,matrix);
+            NeedlemanWunsch aligner = new NeedlemanWunsch((short)-2,(short) 5,(short) 3,(short) 3,(short) 0,matrix);
             
             if (DEBUG){
             	System.out.println("seq lengths: " + bjseq1.seqString().length() + " " + bjseq2.seqString().length());
